+++ CHANGELOG.md
@@ -15,6 +15,9 @@
   allowing type parameters without default values to follow those with
   default values in some type parameter lists. Patch by Alex Waygood,
   backporting a CPython PR by Jelle Zijlstra.
+- It is now disallowed to use a `TypeVar` with a default value after a
+  `TypeVarTuple` in a type parameter list. This matches the CPython
+  implementation of PEP 696 on Python 3.13+.
 - Fix `Protocol` tests on Python 3.13.0a6 and newer. 3.13.0a6 adds a new
   `__static_attributes__` attribute to all classes in Python,
   which broke some assumptions made by the implementation of
@@ -36,6 +39,7 @@
   signifies the return type of the `__(a)exit__` method, matching
   `typing.ContextManager` and `typing.AsyncContextManager` on Python
   3.13+.
+- Backport `types.CapsuleType` from Python 3.13.
 
 # Release 4.11.0 (April 5, 2024)